Refactor games and authors by removing them from helpers
