+++ authors.json
@@ -1,0 +1,1 @@
+[{"first_name":"Yan","last_name":"Dan"}]
