+++ games.json
@@ -1,0 +1,1 @@
+[{"multiplayer":"yes","last_played_at":"2020-02-02","pub_date":"2020/02/02"}]
