+++ helpers.rb
@@ -1,8 +1,6 @@
 require_relative 'book'
-require_relative 'game'
 require_relative 'music_album'
 require_relative 'genre'
-require_relative 'author'
 require 'json'
 
 module Helpers
@@ -33,73 +31,12 @@ module Helpers
     file.close
   end
 
-  # def list_books
-  #   books = read_file('books.json');
-  #   puts 'There are no books in the catalog' if books.empty?
-
-  #   @books.each do |book|
-  #     puts "Publish Data: #{book.publish_date}, cover_state:#{book.cover_state}"
-  #   end
-  # end
-
-  def list_games
-    games = JSON.parse(File.read('games.json'))
-    if games.empty?
-      puts 'No game in Catalog yet. Please add game from menu!'
-    else
-      games.each do |game|
-        puts "Multiplayer: #{game['multiplayer']}"
-        puts "Last played at: #{game['last_played_at']}"
-        puts "Publication date: #{game['pub_date']}"
-        puts ''
-      end
-    end
-  end
-
-  # def list_labels
-
-  # end
-
   def create_music_album_obj(music_album)
     { name: music_album.name, publish_date: music_album.publish_date, on_spotify: music_album.on_spotify }
   end
 
   def create_genre_obj(genre)
     { name: genre.name }
-  end
-
-  def add_author
-    print 'Enter first name: '
-    f_name = gets.chomp
-    print 'Enter last name: '
-    l_name = gets.chomp
-    author = Author.new(f_name, l_name)
-    File.open('authors.json', 'w') { |f| f.write JSON.generate([]) } unless File.exist? 'authors.json'
-    authors = JSON.parse(File.read('authors.json'))
-    authors << { 'first_name' => author.first_name, 'last_name' => author.last_name }
-    File.write('authors.json', JSON.generate(authors))
-  end
-
-  def add_game
-    print 'Is the game multiplayer: '
-    multiplayer = gets.chomp
-    print 'When was it last played [YYYY/MM/DD]: '
-    last_played = gets.chomp
-    print 'What is the publish date of the game: '
-    pub_date = gets.chomp
-    print 'Do you want to add an author? [y] for yes and [n] for no: '
-    author = gets.chomp
-    case author
-    when 'y' then add_author
-    when 'n' then ''
-    end
-    game = Game.new(multiplayer, last_played, pub_date)
-    File.open('games.json', 'w') { |f| f.write JSON.generate([]) } unless File.exist? 'games.json'
-    games = JSON.parse(File.read('games.json'))
-    games << { 'multiplayer' => game.multiplayer, 'last_played_at' => game.last_played_at,
-               'pub_date' => game.publish_date }
-    File.write('games.json', JSON.generate(games))
-    puts 'Game added successfully!'
   end
 
   def load_books
@@ -117,40 +54,4 @@ module Helpers
     end
     save_file('books.json', JSON.generate(data))
   end
-
-  def list_authors
-    authors = JSON.parse(File.read('authors.json'))
-    if authors.empty?
-      puts 'No author in catalog yet!'
-    else
-      authors.each do |author|
-        puts "First name: #{author['first_name']}"
-        puts "Last name: #{author['last_name']}"
-        puts ''
-      end
-    end
-  end
-
-  # def add_book
-  #   puts 'Enter the publisher of the book'
-  #   publisher = gets.chomp
-  #   puts 'Enter the cover state of the book'
-  #   cover_state = gets.chomp
-  #   @books << Book.new(publisher, cover_state)
-  # end
-
-  # def load_books
-  #   data = []
-  #   @books.each do |book|
-  #     data.push(Book.new(book['publisher'], book['cover_state']))
-  #   end
-  #   data
-  # end
-  # def save_books
-  #   data = []
-  #   @books.each do |book|
-  #     data.push({'publisher':book.publisher, 'cover_state': book.cover_state})
-  #   end
-  #   save_file('books.json', JSON.generate(data))
-  # end
 end
